Extract versions test utilities to factory utils.
